# swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/positions/stats": {
      "get": {
        "operationId": "PositionsController_getStats",
        "parameters": [
          {
            "name": "accountId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        },
        "security": [
          {
            "accessToken": []
          }
        ],
        "summary": "Get statistics about positions and orders of the account",
        "tags": [
          "Positions"
        ]
      }
    },
    "/positions/bulk-close": {
      "post": {
        "operationId": "PositionsController_bulkClose",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkClosePositionsDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        },
        "security": [
          {
            "accessToken": []
          }
        ],
        "summary": "Bulk close positions",
        "tags": [
          "Positions"
        ]
      }
    },
    "/positions/{positionId}/close": {
      "post": {
        "operationId": "PositionsController_closePosition",
        "parameters": [
          {
            "name": "positionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClosePositionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        },
        "security": [
          {
            "accessToken": []
          }
        ],
        "summary": "Close position",
        "tags": [
          "Positions"
        ]
      }
    },
    "/positions": {
      "get": {
        "operationId": "PositionsController_getPositions",
        "parameters": [
          {
            "name": "take",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "skip",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 0,
              "default": 0,
              "type": "number"
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "symbol",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "side",
            "required": false,
            "in": "query",
            "description": "Order side: 0 for buy, 1 for sell",
            "schema": {
              "type": "number",
              "enum": [
                0,
                1
              ]
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "schema": {
              "default": "openedAt",
              "type": "string",
              "enum": [
                "openedAt",
                "closedAt",
                "quantity",
                "symbol",
                "side",
                "realizedPnl"
              ]
            }
          },
          {
            "name": "sortDir",
            "required": false,
            "in": "query",
            "schema": {
              "default": "desc",
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "Array of position statuses",
            "schema": {
              "type": "array",
              "items": {
                "type": "number"
              }
            }
          },
          {
            "name": "fromDate",
            "required": false,
            "in": "query",
            "schema": {
              "pattern": "^\\d{4}-\\d{2}-\\d{2}$",
              "type": "string"
            }
          },
          {
            "name": "toDate",
            "required": false,
            "in": "query",
            "schema": {
              "pattern": "^\\d{4}-\\d{2}-\\d{2}$",
              "type": "string"
            }
          },
          {
            "name": "accountId",
            "required": true,
            "in": "query",
            "description": "Trading account ID",
            "schema": {
              "pattern": "^(real|demo)_[0-9]{8}$",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        },
        "security": [
          {
            "accessToken": []
          }
        ],
        "summary": "List positions",
        "tags": [
          "Positions"
        ]
      }
    },
    "/positions/{positionId}": {
      "put": {
        "operationId": "PositionsController_updatePosition",
        "parameters": [
          {
            "name": "positionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePositionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        },
        "security": [
          {
            "accessToken": []
          }
        ],
        "summary": "Update position",
        "tags": [
          "Positions"
        ]
      }
    },
...
  },
  "info": {
    "title": "RWA Backend API",
    "description": "API documentation for the RWA graduation project - Real World Assets Platform",
    "version": "1.0.0",
    "contact": {
      "name": "RWA Team",
      "url": "",
      "email": "support@rwa.com"
    },
    "license": {
      "name": "Proprietary",
      "url": ""
    }
  },
...
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local Development"
    }
  ],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Enter JWT token"
      },
      "accessToken": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "refreshToken": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "apiKey": {
        "type": "apiKey",
        "name": "apiKey",
        "in": "header"
      }
    },
    "schemas": {
...
      "BulkClosePositionsDto": {
        "type": "object",
        "properties": {
          "accountId": {
            "type": "string",
            "pattern": "^(real|demo)_[0-9]{8}$",
            "description": "Trading account ID"
          },
          "positionIds": {
            "minItems": 1,
            "description": "Array of position IDs to close",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "accountId",
          "positionIds"
        ]
      },
      "ClosePositionDto": {
        "type": "object",
        "properties": {
          "accountId": {
            "type": "string",
            "pattern": "^(real|demo)_[0-9]{8}$",
            "description": "Trading account ID"
          },
          "quantity": {
            "type": "object",
            "description": "Close quantity (partial/full)"
          }
        },
        "required": [
          "accountId",
          "quantity"
        ]
      },
      "UpdatePositionDto": {
        "type": "object",
        "properties": {
          "accountId": {
            "type": "string",
            "pattern": "^(real|demo)_[0-9]{8}$",
            "description": "Trading account ID"
          },
          "stopLossPrice": {
            "type": "object",
            "description": "New stop loss price. Optional, set to null to remove"
          },
          "takeProfitPrice": {
            "type": "object",
            "description": "New take profit price. Optional, set to null to remove"
          }
        },
        "required": [
          "accountId"
        ]
      },
      "ValidationErrorItem": {
        "type": "object",
        "required": [
          "field",
          "message"
        ],
        "properties": {
          "field": {
            "type": "string"
          },
          "code": {
            "type": "string",
            "description": "Machine-friendly field-level error code."
          },
          "message": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "ErrorResponse": {
        "type": "object",
        "required": [
          "success",
          "code",
          "t"
        ],
        "properties": {
          "success": {
            "type": "boolean",
            "enum": [
              false
            ]
          },
          "code": {
            "type": "string",
            "enum": [
              "internal-error",
              "not-found",
              "bad-request",
              "validation-error",
              "rate-limit-exceeded",
              "duplicate-request",
              "unauthorized",
              "forbidden",
              "token-expired",
              "asset-not-found",
              "invalid-token",
              "token-not-found",
              "session-not-found",
              "session-revoked",
              "session-expired",
              "user-not-found",
              "user-not-active",
              "user-already-exists",
              "invalid-credentials",
              "reference-code-not-found",
              "reference-code-invalid",
              "reference-code-already-set",
              "password-attempt-limit-exceeded",
              "otp-not-found",
              "invalid-otp",
              "invalid-otp-token",
              "mfa-already-enabled",
              "mfa-not-enabled",
              "invalid-mfa-code",
              "challenge-required",
              "challenge-not-found",
              "challenge-expired",
              "challenge-already-used",
              "challenge-invalid-factor",
              "challenge-payload-mismatch",
              "account-not-found",
              "account-not-active",
              "position-not-found",
              "position-already-closed",
              "feedback-not-found",
              "post-not-found",
              "post-unauthorized",
              "comment-not-found",
              "comment-unauthorized",
              "support-ticket-not-found",
              "support-unauthorized",
              "insufficient-balance",
              "transfer-to-same-account",
              "transfer-not-allowed-with-open-position",
              "external-api-error",
              "invalid-quantity",
              "price-not-found"
            ]
          },
          "message": {
            "type": "string",
            "description": "Human-readable error summary for clients/logging."
          },
          "t": {
            "type": "string",
            "format": "date-time"
          },
          "errors": {
            "description": "Detailed error information by field, domain rule, or external provider.",
            "oneOf": [
              {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/ValidationErrorItem"
                }
              },
              {
                "type": "object",
                "additionalProperties": true
              }
            ]
          },
          "traceId": {
            "type": "string",
            "description": "Request trace id for debugging across services."
          }
        },
        "additionalProperties": false
      },
...
    },
    "responses": {
      "BadRequest": {
        "description": "Malformed request payload or failed domain validation.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Authentication is missing, invalid, revoked, or expired.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "Forbidden": {
        "description": "Authenticated but not allowed to access or mutate this resource.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "NotFound": {
        "description": "Requested resource was not found.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "TooManyRequests": {
        "description": "Rate limit exceeded or too many OTP/password attempts.",
        "headers": {
          "Retry-After": {
            "description": "Seconds to wait before retrying.",
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "InternalServerError": {
        "description": "Unexpected server error or upstream dependency failure.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      }
    }
  }
}